Added quiz functionality
You can now do a quiz with radio buttons and get a score. Currently the right answer is button number 1 on all questions. There is also a bug that the last answer does not count because viewpager already loads the last page before the user can press the button. Will probbably ask professor about this.